feat: implement routing and main application structure with About page
