+++ src/components/HeaderArea.tsx
@@ -1,22 +1,29 @@
+import { NavLink } from 'react-router'
+
 function HeaderArea() {
   return (
     <header>
       <nav className="top-0 left-0 mb-10 flex h-15 w-full bg-white text-black shadow-sm">
         <h1 className="my-auto mr-5 ml-5 text-2xl font-bold">
-          <a href="#">
-            <img
-              src="/favicon.svg"
-              alt="logo"
-              width="20"
-              height="20"
-              className="mr-1 hidden md:inline"
-            />
-            PictRider
-          </a>
+          <img
+            src="/favicon.svg"
+            alt="logo"
+            width="20"
+            height="20"
+            className="mr-1 hidden md:inline"
+          />
+          PictRider
         </h1>
         <ul className="mt-auto mb-auto flex space-x-6 text-lg font-bold">
           <li>
-            <a href="#">About</a>
+            <NavLink to="/" end>
+              Home
+            </NavLink>
+          </li>
+          <li>
+            <NavLink to="/about" end>
+              About
+            </NavLink>
           </li>
           <li>
             <a
