+++ src/pages/index.tsx
@@ -1,0 +1,2 @@
+export { default as AppMain } from './AppMain'
+export { default as About } from './About'
